pyproject.toml
[project]
name = "ght-qcxr"
version = "0.1.0"
description = "Northwestern Global Health Technologies CXR Quality Assessor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "joblib>=1.5.3",
    "medmnist>=3.0.2",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.0",
    "protobuf>=6.33.5",
    "pydicom>=3.0.1",
    "pyjpegls>=1.5.1",
    "pylibjpeg[all]>=2.1.0",
    "pyorthanc>=1.22.1",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "sentencepiece>=0.2.1",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "torchxrayvision>=1.4.0",
    "transformers>=5.0.0",
    "ultralytics>=8.4.14",
]
